refactor: extract S3 upload into a dial9-s3 crate, drop aws from core
