+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "dial9-macro",
     "dial9-viewer",
     "dial9",
+    "dial9-s3",
     "perf-self-profile",
     "examples/metrics-service",
     "examples/simple-local",
@@ -24,6 +25,7 @@ repository = "https://github.com/dial9-rs/dial9-tokio-telemetry"
 [workspace.dependencies]
 libc = "0.2"
 dial9-core = { version = "0.4.0-alpha.1", path = "dial9-core" }
+dial9-s3 = { version = "0.4.0-alpha.1", path = "dial9-s3" }
 dial9-perf-self-profile = { version = "0.4.4", path = "perf-self-profile" }
 dial9-trace-format = { version = "0.4.1", path = "dial9-trace-format" }
 dial9-trace-format-derive = { version = "0.4.1", path = "dial9-trace-format-derive" }
