+++ dial9-core/Cargo.toml
@@ -27,22 +27,9 @@ metrique = "0.1.23"
 metrique-timesource = "0.1"
 shuttle = { version = "0.9.1", optional = true }
 tracing = "0.1.44"
-aws-sdk-s3-transfer-manager = { version = "0.1.3", optional = true, default-features = false }
-aws-sdk-s3 = { version = "1", default-features = false, optional = true, features = [
-    "behavior-version-latest",
-    "sigv4a",
-    "rt-tokio",
-    "default-https-client",
-] }
-aws-config = { version = "1", optional = true, default-features = false }
-time = { version = "0.3", features = ["formatting", "macros"], optional = true }
-hostname = { version = "0.4", optional = true }
-serde = { version = "1", features = ["derive"], optional = true }
-serde_json = { version = "1", optional = true }
 
 [dev-dependencies]
 assert2 = { workspace = true }
-async-trait = "0.1.89"
 dial9-trace-format = { workspace = true, features = ["serde-deserialize"] }
 metrique-timesource = { version = "0.1", features = [
     "custom-timesource",
@@ -53,10 +40,13 @@ metrique-writer = { version = "0.1", features = ["test-util"] }
 proptest = "1"
 serde = { version = "1", features = ["derive"] }
 tempfile = "3"
-tokio = { version = "1.52.0", features = ["rt", "macros", "time", "test-util"] }
-s3s = "0.13.0"
-s3s-fs = "0.13.0"
-s3s-aws = "0.13.0"
+tokio = { version = "1.52.0", features = [
+    "rt",
+    "rt-multi-thread",
+    "macros",
+    "time",
+    "test-util",
+] }
 
 [features]
 _shuttle = ["dep:shuttle", "pipeline"]
@@ -72,15 +62,4 @@ pipeline = [
     "dep:tokio-util",
     "dep:futures-util",
     "dep:flate2",
-]
-## Built-in S3 upload stage. Implies the pipeline runner, pulls the aws SDK.
-pipeline-s3 = [
-    "pipeline",
-    "dep:aws-sdk-s3-transfer-manager",
-    "dep:aws-sdk-s3",
-    "dep:aws-config",
-    "dep:time",
-    "dep:hostname",
-    "dep:serde",
-    "dep:serde_json",
 ]
