+++ dial9-s3/Cargo.toml
@@ -1,0 +1,40 @@
+[package]
+name = "dial9-s3"
+version = "0.4.0-alpha.1"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "S3 upload SegmentProcessor for dial9-core's pipeline"
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ["cfg(shuttle)"] }
+
+[dependencies]
+dial9-core = { workspace = true, features = ["pipeline"] }
+bon = "3"
+bytes = "1"
+tokio = { version = "1", features = ["fs"], default-features = false }
+tracing = "0.1.44"
+aws-sdk-s3-transfer-manager = { version = "0.1.3", default-features = false }
+aws-sdk-s3 = { version = "1", default-features = false, features = [
+    "behavior-version-latest",
+    "sigv4a",
+    "rt-tokio",
+    "default-https-client",
+] }
+aws-config = { version = "1", default-features = false }
+time = { version = "0.3", features = ["formatting", "macros"] }
+hostname = "0.4"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+
+[dev-dependencies]
+dial9-core = { workspace = true, features = ["pipeline", "test-util"] }
+assert2 = { workspace = true }
+async-trait = "0.1.89"
+flate2 = "1"
+tempfile = "3"
+tokio = { version = "1.52.0", features = ["rt", "macros", "time", "test-util"] }
+s3s = "0.13.0"
+s3s-fs = "0.13.0"
+s3s-aws = "0.13.0"
