+++ dial9-tokio-telemetry/Cargo.toml
@@ -72,6 +72,7 @@ metrique-timesource = "0.1"
 dial9-macro = { workspace = true }
 shuttle = { version = "0.9.1", optional = true }
 dial9-core = { workspace = true, features = ["pipeline"] }
+dial9-s3 = { workspace = true, optional = true }
 
 [features]
 analysis = []
@@ -96,7 +97,7 @@ tracing-layer = ["dep:tracing-subscriber"]
 ## Sampled memory allocation profiling via ring buffers.
 memory-profiling = ["dep:dial9-perf-self-profile", "dep:scc"]
 worker-s3 = [
-    "dial9-core/pipeline-s3",
+    "dep:dial9-s3",
     "dep:aws-sdk-s3-transfer-manager",
     "dep:aws-sdk-s3",
     "dep:aws-config",
